Add files via upload
In this script, I calculatet the mean percentage of state employment for non-democratic and democratic countries. The code also calculates the percentage of the sample that depends on state-provided assistance. Based on calculations, I generate a bar chart that compares the mean percentage of state employment in non-democratic and democratic countries. Next, I create a scatter plot that shows the relationship between the average percentage of people receiving state assistance and the average percentage of state employment for each country in Post-Communist Europe and Central Asia.